features/rbac/ports: consolidate PolicyCondition + ConditionOperator into canonical types (Issue #1278)
Three packages in features/rbac each declared their own PolicyCondition struct,
causing a type-duplication cycle. Two packages (jit and memory) also defined
ConditionOperator as a plain string rather than a typed constant, reducing
type-safety at construction sites.

This commit introduces features/rbac/ports as the canonical home for shared
RBAC port types (part of Epic #738 — resolve duplicate-type cycle workarounds).

What changed:
- Create features/rbac/ports/types.go: declares canonical PolicyCondition struct
  (Field string, Operator ConditionOperator, Value interface{}, ValueType string)
  and ConditionOperator type with all 8 constants
- Create features/rbac/ports/types_test.go: compile-time assertions covering all
  4 struct fields and all 8 ConditionOperator constants
- features/rbac/memory/store.go: delete local PolicyCondition struct (Attribute
  field, untyped Operator); add ports import; alias to ports.PolicyCondition
- features/rbac/jit/interfaces.go: delete local PolicyCondition struct (untyped
  Operator); add ports import; alias to ports.PolicyCondition
- features/rbac/zerotrust/types.go: delete PolicyCondition struct and
  ConditionOperator type+8 constants; add ports import; alias both types

zerotrust/policy_language.go construction sites at lines 618 and 716 use
ConditionOperator(condSpec.Operator) string-to-typed casts — these continue to
compile correctly because the alias preserves the identical underlying string type.

Fixes #1278
